Add a features page (#126)
* Quick first draft of a features (comparison) table

Just want to get something going. It's a basic markdown table. Each row lists a feature. The name is a link to the respective place in the docs. The other two columns are to indicate support in the application and support in the QT application. Very simple.

If we like this format, we can properly gather the table content.

Closes #125

* Added more features and restructuring

After realizing that this will be a very long list, I split things up into multiple tables.

* Adding some more info

Just going to keep iterating until we're there.

* Lots more updates