Light page restructuring (#146)
* Light page restructuring

Moving some of the designs that are beyond the parity point with the QT wallet into a "Milestones 2.x" section. Makes for a clearer separation of exploratory features vs. the core of the application.

Note that this is not a 100% clean separation across pages. Within a feature, there might be multiple versions to get implemented iteratively. But at least this one moves some of the big ones into a dedicated category. We'll continue to tweak this as we go.

Also sneaking in some minor fixes here.
- Adding missing favicons
- Removing web manifest as its not needed
- Fixing a bug in the picture template so it doesn't render empty image sources

Also added a rake file to automatically detect broken links and URLs. And fixed a few of them.

* Moving multi-key to the 2.x milestone

Also:
- Updating the roadmap page with the 2.x milestones
- Moving block status page out of the milestones folder (not sure why it's in there)
- Adding list of settings sub-pages to the settings main page.